System Agent set the initial AID in agent's map, be aware that DA & PA have setup commented due to trials purposes
When initial game is initializated only the agent types are set to the cells with mobile agents but their AID is unknown as they are not created inside the container yet.
This is not possible to generate a container in generate initial game, otherwise, this will be the best place to create all agents and set the AIDs.
Now the System Agent sets all AIDs in the game.